requirements.txt
python-telegram-bot
python-dotenv
httpx
Flask==3.1.0
pydantic